Add restricted flag to triggers
